fix: prevent deletion of active events with registrations

+++ src/main/java/com/feponiel/swiftpass/domain/application/usecases/DeleteEventUseCase.java
@@ -1,11 +1,17 @@
 package com.feponiel.swiftpass.domain.application.usecases;
 
+import java.util.List;
 import java.util.UUID;
 
 import org.springframework.stereotype.Service;
 
 import com.feponiel.swiftpass.domain.application.repositories.EventsRepository;
+import com.feponiel.swiftpass.domain.application.repositories.RegistrationsRepository;
+import com.feponiel.swiftpass.domain.application.usecases.exceptions.ActiveEventDeletionNotAllowedException;
 import com.feponiel.swiftpass.domain.application.usecases.exceptions.EventNotFoundException;
+import com.feponiel.swiftpass.domain.business.entities.Event;
+import com.feponiel.swiftpass.domain.business.entities.Registration;
+import com.feponiel.swiftpass.domain.business.valueobjects.EventStatus;
 
 import lombok.RequiredArgsConstructor;
 
@@ -13,10 +19,17 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class DeleteEventUseCase {
   private final EventsRepository eventsRepository;
+  private final RegistrationsRepository registrationsRepository;
 
   public void execute(UUID eventId) {
-    this.eventsRepository.findById(eventId)
+    Event event = this.eventsRepository.findById(eventId)
       .orElseThrow(EventNotFoundException::new);
+
+    List<Registration> eventRegistrationList = registrationsRepository.listAllByEventIdAndPaymentStatus(eventId, null);
+
+    if (eventRegistrationList.size() > 0 && event.getStatus() != EventStatus.CANCELED) {
+      throw new ActiveEventDeletionNotAllowedException();
+    }
 
     this.eventsRepository.deleteById(eventId);
   }
